requirements.txt
snowflake-connector-python
json5
pysqlite3-binary
crewai[tools]
langchain-openai
python-dotenv
streamlit
gevent
